Moved Statoil independent ecl_file tests.
